youtube_txt/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1685243452975/work
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
black @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_29yzn6blpk/croot/black_1680737263429/work
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_croot-quwchbn8/click_1646123324461/work
cryptography==41.0.2
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1686074432687/work
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1683904088844/work
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
firebase-admin==6.2.0
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.3
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==0.4.3
openai==0.27.8
packaging @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e946luvhc3/croot/packaging_1678965323926/work
pathspec @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f954ebvbxx/croot/pathspec_1674681573969/work
platformdirs @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f7wx6m2jsp/croots/recipe/platformdirs_1662711384790/work
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1681817562700/work
tomli @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d0e5ffbf-5cf1-45be-8693-c5dff8108a2awhthtjlq/croots/recipe/tomli_1657175508477/work
tqdm==4.65.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685704949284/work
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.15.0
